Update call method on the MoveZeros class
To moves all of the zeros of an array to the end while preserving
 the order of the other elements, the partition method has been
used to separate the 0 and group them into a second array. This new
array is then pushed at the end using the reverse method prior being
compressed again into one single array using the flatten method.